pyproject.toml
[tool.poetry]
name = "hw-web-11"
version = "0.0.1"
description = "RST API Contacts"
authors = ["s.kravtsov <s.kravtsov@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
SQLAlchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"
Jinja2 = "^3.1.3"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.38.0"

[tool.poetry.dev-dependencies]
Sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
